+++ client/web-admin-2/src/components/routers/Volunteer.js
@@ -1,29 +1,10 @@
 import React, { useEffect } from 'react'
 import { Switch, Route, Redirect } from 'react-router-dom'
-import { useMyJurisdictions, useWipJurisdiction } from 'store/selectors'
+import { useMyJurisdictions } from 'store/selectors'
 import useVolunteerActions from 'store/actions/volunteer'
 import Layout from 'components/core/Layout'
 import MyJurisdictions from 'components/MyJurisdictions'
-import JurisdictionDetails from 'components/JurisdictionDetails'
-import EditLocation from 'components/EditLocation'
-
-// ensure that the wipJurisdiction is loaded before any subroutes are rendered
-const WipJurisdictionLoader = ({ match: { params: { jid }, path } }) => {
-  const { getWipJurisdiction } = useVolunteerActions()
-  const wipJurisdiction = useWipJurisdiction(jid)
-
-  useEffect(() => {
-    if (!wipJurisdiction) getWipJurisdiction(jid)
-  }, [jid, wipJurisdiction, getWipJurisdiction])
-
-  if (!wipJurisdiction) return null
-  return (
-    <Switch>
-      <Route exact path={path} component={JurisdictionDetails} />
-      <Route path={`${path}/locations/:lid`} component={EditLocation} />
-    </Switch>
-  )
-}
+import JurisdictionEditor from 'components/JurisdictionEditor'
 
 const Volunteer = () => {
   const myJurisdictions = useMyJurisdictions()
@@ -38,7 +19,7 @@ const Volunteer = () => {
     <Layout>
       <Switch>
         <Route exact path="/jurisdictions" component={MyJurisdictions} />
-        <Route path="/jurisdictions/:jid" component={WipJurisdictionLoader} />
+        <Route path="/jurisdictions/:jid" component={JurisdictionEditor} />
         <Redirect to="/jurisdictions" />
       </Switch>
     </Layout>
